# api-gateway/swagger.yaml
openapi: 3.0.0
info:
  title: LinkedIn AgenticAI API
  version: 1.0.0
  description: |
    Complete distributed system backend API documentation for Data 236 Project.

    **Agentic AI (FastAPI):** Paths under `/ai/*` match **Uvicorn** on port **8001** exactly.
    Use this gateway (`/api` prefix) for **Try it out** — e.g. `POST /api/ai/tasks/submit` → `http://localhost:8001/ai/tasks/submit`.
    **WebSocket:** `ws://localhost:4000/api/ai/ws/ai/tasks/{task_id}` (proxied to FastAPI `/ws/ai/tasks/{task_id}`).
    Canonical interactive docs for the Python service alone: `http://localhost:8001/docs`.
servers:
  - url: http://localhost:4000/api
    description: Local API Gateway
paths:
...
  /jobs/create:
    post:
      tags: [Job Service]
      summary: Create a new job posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                company:
                  type: string
                location:
                  type: string
                salary:
                  type: string
                type:
                  type: string
                description:
                  type: string
                skills:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Created
...